+++ CountlyConfig.m
@@ -6,6 +6,10 @@
 
 #import "CountlyCommon.h"
 
+@interface CountlyConfig ()
+@property (nonatomic) NSMutableArray<RCDownloadCallback> *remoteConfigGlobalCallbacks;
+@end
+
 @implementation CountlyConfig
 
 //NOTE: Countly features
@@ -62,6 +66,17 @@ CLYDeviceIDType const CLYDeviceIDTypeNSUUID     = @"CLYDeviceIDTypeNSUUID";
     }
 
     return self;
+}
+
+-(void)remoteConfigRegisterGlobalCallback:(RCDownloadCallback) callback
+{
+    [self.remoteConfigGlobalCallbacks addObject:callback];
+}
+
+
+- (NSMutableArray<RCDownloadCallback> *) getRemoteConfigGlobalCallbacks
+{
+    return self.remoteConfigGlobalCallbacks;
 }
 
 @end