+++ CountlyConsentManager.m
@@ -433,11 +433,12 @@ CLYConsent const CLYConsentRemoteConfig         = @"remote-config";
     {
         CLY_LOG_D(@"Consent for RemoteConfig is given.");
 
-        [CountlyRemoteConfig.sharedInstance startRemoteConfig];
+        [CountlyRemoteConfigInternal.sharedInstance downloadRemoteConfigAutomatically];
     }
     else
     {
         CLY_LOG_D(@"Consent for RemoteConfig is cancelled.");
+        [CountlyRemoteConfigInternal.sharedInstance clearCachedRemoteConfig:YES];
     }
 }