+++ CountlyPersistency.m
@@ -495,16 +495,17 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
 
 - (NSDictionary *)retrieveRemoteConfig
 {
-    NSDictionary* remoteConfig = [NSUserDefaults.standardUserDefaults objectForKey:kCountlyRemoteConfigPersistencyKey];
+    NSData* data = [NSUserDefaults.standardUserDefaults objectForKey:kCountlyRemoteConfigPersistencyKey];
+    NSDictionary* remoteConfig = [NSKeyedUnarchiver unarchiveObjectWithData:data];
     if (!remoteConfig)
         remoteConfig = NSDictionary.new;
-
+    
     return remoteConfig;
 }
 
 - (void)storeRemoteConfig:(NSDictionary *)remoteConfig
 {
-    [NSUserDefaults.standardUserDefaults setObject:remoteConfig forKey:kCountlyRemoteConfigPersistencyKey];
+    [NSUserDefaults.standardUserDefaults setObject:[NSKeyedArchiver archivedDataWithRootObject:remoteConfig] forKey:kCountlyRemoteConfigPersistencyKey];
     [NSUserDefaults.standardUserDefaults synchronize];
 }