+++ CountlyRemoteConfig.m
@@ -1,18 +1,12 @@
-// CountlyLocationManager.h
 //
-// This code is provided under the MIT License.
+//  CountlyRemoteConfig.m
+//  CountlyTestApp-iOS
 //
-// Please visit www.count.ly for more information.
+//  Created by Muhammad Junaid Akram on 07/06/2023.
+//  Copyright © 2023 Countly. All rights reserved.
+//
 
 #import "CountlyCommon.h"
-
-NSString* const kCountlyRCKeyFetchRemoteConfig  = @"fetch_remote_config";
-NSString* const kCountlyRCKeyKeys               = @"keys";
-NSString* const kCountlyRCKeyOmitKeys           = @"omit_keys";
-
-@interface CountlyRemoteConfig ()
-@property (nonatomic) NSDictionary* cachedRemoteConfig;
-@end
 
 @implementation CountlyRemoteConfig
 
@@ -20,7 +14,7 @@ NSString* const kCountlyRCKeyOmitKeys           = @"omit_keys";
 {
     if (!CountlyCommon.sharedInstance.hasStarted)
         return nil;
-
+    
     static CountlyRemoteConfig* s_sharedInstance = nil;
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{s_sharedInstance = self.new;});
@@ -31,192 +25,100 @@ NSString* const kCountlyRCKeyOmitKeys           = @"omit_keys";
 {
     if (self = [super init])
     {
-        self.cachedRemoteConfig = [CountlyPersistency.sharedInstance retrieveRemoteConfig];
+        
     }
-
+    
     return self;
 }
 
-#pragma mark ---
-
-- (void)startRemoteConfig
+- (NSDictionary *)testingGetAllVariants
 {
-    if (!self.isEnabledOnInitialConfig)
-        return;
-
-    if (!CountlyConsentManager.sharedInstance.consentForRemoteConfig)
-        return;
-
-    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
-        return;
-
-    CLY_LOG_D(@"Fetching remote config on start...");
-
-    [self fetchRemoteConfigForKeys:nil omitKeys:nil completionHandler:^(NSDictionary *remoteConfig, NSError *error)
-    {
-        if (!error)
-        {
-            CLY_LOG_D(@"Fetching remote config on start is successful. \n%@", remoteConfig);
-
-            self.cachedRemoteConfig = remoteConfig;
-            [CountlyPersistency.sharedInstance storeRemoteConfig:self.cachedRemoteConfig];
-        }
-        else
-        {
-            CLY_LOG_W(@"Fetching remote config on start failed: %@", error);
-        }
-
-        if (self.remoteConfigCompletionHandler)
-            self.remoteConfigCompletionHandler(error);
-    }];
+    CLY_LOG_I(@"%s", __FUNCTION__);
+    
+    return [CountlyRemoteConfigInternal.sharedInstance testingGetAllVariants];
 }
 
-- (void)updateRemoteConfigForKeys:(NSArray *)keys omitKeys:(NSArray *)omitKeys completionHandler:(void (^)(NSError * error))completionHandler
-{
-    if (!CountlyConsentManager.sharedInstance.consentForRemoteConfig)
-        return;
-
-    if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
-        return;
-
-    CLY_LOG_D(@"Fetching remote config manually...");
-
-    [self fetchRemoteConfigForKeys:keys omitKeys:omitKeys completionHandler:^(NSDictionary *remoteConfig, NSError *error)
-    {
-        if (!error)
-        {
-            CLY_LOG_D(@"Fetching remote config manually is successful. \n%@", remoteConfig);
-
-            if (!keys && !omitKeys)
-            {
-                self.cachedRemoteConfig = remoteConfig;
-            }
-            else
-            {
-                NSMutableDictionary* partiallyUpdatedRemoteConfig = self.cachedRemoteConfig.mutableCopy;
-                [partiallyUpdatedRemoteConfig addEntriesFromDictionary:remoteConfig];
-                self.cachedRemoteConfig = [NSDictionary dictionaryWithDictionary:partiallyUpdatedRemoteConfig];
-            }
-
-            [CountlyPersistency.sharedInstance storeRemoteConfig:self.cachedRemoteConfig];
-        }
-        else
-        {
-            CLY_LOG_W(@"Fetching remote config manually failed: %@", error);
-        }
-
-        if (completionHandler)
-            completionHandler(error);
-    }];
+- (NSArray *)testingGetVariantsForKey:(NSString *)key {
+    
+    CLY_LOG_I(@"%s %@", __FUNCTION__, key);
+    
+    return [CountlyRemoteConfigInternal.sharedInstance testingGetVariantsForKey:key];
 }
 
-- (id)remoteConfigValueForKey:(NSString *)key
-{
-    return self.cachedRemoteConfig[key];
+- (void)testingEnrollIntoVariant:(NSString *)key variantName:(NSString *)variantName completionHandler:(RCVariantCallback)completionHandler {
+    
+    CLY_LOG_I(@"%s %@ %@ %@", __FUNCTION__, key, variantName , completionHandler);
+    
+    [CountlyRemoteConfigInternal.sharedInstance testingEnrollIntoVariant:key variantName:variantName completionHandler:completionHandler];
 }
 
-- (void)clearCachedRemoteConfig
+- (void)testingDownloadVariantInformation:(RCVariantCallback)completionHandler
 {
-    self.cachedRemoteConfig = nil;
-    [CountlyPersistency.sharedInstance storeRemoteConfig:self.cachedRemoteConfig];
+    CLY_LOG_I(@"%s %@", __FUNCTION__, completionHandler);
+    
+    [CountlyRemoteConfigInternal.sharedInstance testingDownloadAllVariants:nil completionHandler:completionHandler];
 }
 
-#pragma mark ---
-
-- (void)fetchRemoteConfigForKeys:(NSArray *)keys omitKeys:(NSArray *)omitKeys completionHandler:(void (^)(NSDictionary* remoteConfig, NSError * error))completionHandler
+- (CountlyRCData *)getValue:(NSString *)key
 {
-    if (!CountlyServerConfig.sharedInstance.networkingEnabled)
-    {
-        CLY_LOG_D(@"'fetchRemoteConfigForKeys' is aborted: SDK Networking is disabled from server config!");
-        return;
-    }
-    if (!completionHandler)
-        return;
-
-    NSURLRequest* request = [self remoteConfigRequestForKeys:keys omitKeys:omitKeys];
-    NSURLSessionTask* task = [NSURLSession.sharedSession dataTaskWithRequest:request completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
-    {
-        NSDictionary* remoteConfig = nil;
-
-        if (!error)
-        {
-            remoteConfig = [NSJSONSerialization JSONObjectWithData:data options:0 error:&error];
-        }
-
-        if (!error)
-        {
-            if (((NSHTTPURLResponse*)response).statusCode != 200)
-            {
-                NSMutableDictionary* userInfo = remoteConfig.mutableCopy;
-                userInfo[NSLocalizedDescriptionKey] = @"Remote config general API error";
-                error = [NSError errorWithDomain:kCountlyErrorDomain code:CLYErrorRemoteConfigGeneralAPIError userInfo:userInfo];
-            }
-        }
-
-        if (error)
-        {
-            CLY_LOG_D(@"Remote Config Request <%p> failed!\nError: %@", request, error);
-
-            dispatch_async(dispatch_get_main_queue(), ^
-            {
-                completionHandler(nil, error);
-            });
-
-            return;
-        }
-
-        CLY_LOG_D(@"Remote Config Request <%p> successfully completed.", request);
-
-        dispatch_async(dispatch_get_main_queue(), ^
-        {
-            completionHandler(remoteConfig, nil);
-        });
-    }];
-
-    [task resume];
-
-    CLY_LOG_D(@"Remote Config Request <%p> started:\n[%@] %@", (id)request, request.HTTPMethod, request.URL.absoluteString);
+    CLY_LOG_I(@"%s %@", __FUNCTION__, key);
+    return [CountlyRemoteConfigInternal.sharedInstance getValue:key];
 }
 
-- (NSURLRequest *)remoteConfigRequestForKeys:(NSArray *)keys omitKeys:(NSArray *)omitKeys
+- (NSDictionary<NSString*, CountlyRCData *> *)getAllValues
 {
-    NSString* queryString = [CountlyConnectionManager.sharedInstance queryEssentials];
-
-    queryString = [queryString stringByAppendingFormat:@"&%@=%@", kCountlyQSKeyMethod, kCountlyRCKeyFetchRemoteConfig];
-
-    if (keys)
-    {
-        queryString = [queryString stringByAppendingFormat:@"&%@=%@", kCountlyRCKeyKeys, [keys cly_JSONify]];
-    }
-    else if (omitKeys)
-    {
-        queryString = [queryString stringByAppendingFormat:@"&%@=%@", kCountlyRCKeyOmitKeys, [omitKeys cly_JSONify]];
-    }
-
-    if (CountlyConsentManager.sharedInstance.consentForSessions)
-    {
-        queryString = [queryString stringByAppendingFormat:@"&%@=%@", kCountlyQSKeyMetrics, [CountlyDeviceInfo metrics]];
-    }
-
-    queryString = [CountlyConnectionManager.sharedInstance appendChecksum:queryString];
-
-    NSString* serverOutputSDKEndpoint = [CountlyConnectionManager.sharedInstance.host stringByAppendingFormat:@"%@%@",
-                                         kCountlyEndpointO,
-                                         kCountlyEndpointSDK];
-
-    if (CountlyConnectionManager.sharedInstance.alwaysUsePOST)
-    {
-        NSMutableURLRequest* request = [NSMutableURLRequest requestWithURL:[NSURL URLWithString:serverOutputSDKEndpoint]];
-        request.HTTPMethod = @"POST";
-        request.HTTPBody = [queryString cly_dataUTF8];
-        return request.copy;
-    }
-    else
-    {
-        NSString* withQueryString = [serverOutputSDKEndpoint stringByAppendingFormat:@"?%@", queryString];
-        NSURLRequest* request = [NSURLRequest requestWithURL:[NSURL URLWithString:withQueryString]];
-        return request;
-    }
+    CLY_LOG_I(@"%s", __FUNCTION__);
+    return [CountlyRemoteConfigInternal.sharedInstance getAllValues];
 }
+
+- (void)enrollIntoABTestsForKeys:(NSArray *)keys
+{
+    CLY_LOG_I(@"%s %@", __FUNCTION__, keys);
+    [CountlyRemoteConfigInternal.sharedInstance enrollIntoABTestsForKeys:keys];
+}
+
+- (void)exitABTestsForKeys:(NSArray *)keys
+{
+    CLY_LOG_I(@"%s %@", __FUNCTION__, keys);
+    [CountlyRemoteConfigInternal.sharedInstance exitABTestsForKeys:keys];
+}
+
+-(void)registerDownloadCallback:(RCDownloadCallback) callback
+{
+    CLY_LOG_I(@"%s %@", __FUNCTION__, callback);
+    [CountlyRemoteConfigInternal.sharedInstance registerDownloadCallback:callback];
+}
+
+-(void)removeDownloadCallback:(RCDownloadCallback) callback
+{
+    CLY_LOG_I(@"%s %@", __FUNCTION__, callback);
+    [CountlyRemoteConfigInternal.sharedInstance removeDownloadCallback:callback];
+}
+
+- (void)downloadKeys:(RCDownloadCallback)completionHandler
+{
+    CLY_LOG_I(@"%s %@", __FUNCTION__, completionHandler);
+    [CountlyRemoteConfigInternal.sharedInstance downloadValuesForKeys:nil omitKeys:nil completionHandler:completionHandler];
+    
+}
+
+- (void)downloadSpecificKeys:(NSArray *)keys completionHandler:(RCDownloadCallback)completionHandler
+{
+    CLY_LOG_I(@"%s %@ %@", __FUNCTION__, keys, completionHandler);
+    [CountlyRemoteConfigInternal.sharedInstance downloadValuesForKeys:keys omitKeys:nil completionHandler:completionHandler];
+    
+}
+
+- (void)downloadOmittingKeys:(NSArray *)omitKeys completionHandler:(RCDownloadCallback)completionHandler
+{
+    CLY_LOG_I(@"%s %@ %@", __FUNCTION__, omitKeys, completionHandler);
+    [CountlyRemoteConfigInternal.sharedInstance downloadValuesForKeys:nil omitKeys:omitKeys completionHandler:completionHandler];
+}
+
+- (void)clearAll
+{
+    [CountlyRemoteConfigInternal.sharedInstance clearCachedRemoteConfig:YES];
+}
+
 
 @end